Refactor context card modules
