+++ tests/test-correctness-phase2.js
@@ -116,6 +116,9 @@ const pwaAppShellAssets = [
   '/js/category-page-view.js',
   '/js/category-view-renderers.js',
   '/js/category-customization.js',
+  '/js/context-cards.js',
+  '/js/context-card-summaries.js',
+  '/js/context-card-editor-ui.js',
   '/js/commit-hash.js',
   '/js/focus-card.js',
   '/js/onboarding-view.js',
