+++ tests/test-emf.js
@@ -143,6 +143,7 @@ assert('45d2. main.js starts the startup orchestrator', mainSrc.includes('startA
 assert('45e. main.js no longer owns the EMF function list', !mainSrc.includes('const _emfFns'));
 const navSrc = read('js/nav.js');
 const contextCardsSrc = read('js/context-cards.js');
+const contextCardSummariesSrc = read('js/context-card-summaries.js');
 const contextProfileCss = read('css/context-profile.css');
 const lensPagesSrc = read('js/lens-pages.js');
 assert('45f. Sidebar exposes EMF assessment under Analysis tools', (() => {
@@ -155,9 +156,9 @@ assert('45g. EMF sidebar entry is available before assessment data exists',
   navSrc.includes('emfAssessmentCount > 0 ? emfAssessmentCount : null') &&
   !navSrc.includes("label: 'EMF',"));
 assert('45h. Environment context card counts saved EMF assessments',
-  contextCardsSrc.includes("key === 'environment'") && contextCardsSrc.includes('getEMFAssessments().length > 0'));
+  contextCardSummariesSrc.includes("key === 'environment'") && contextCardSummariesSrc.includes('getEMFAssessments().length > 0'));
 assert('45i. Environment editor uses the EMF assessment launcher',
-  contextCardsSrc.includes('renderEMFAssessmentLauncher({ inModal: true') && contextCardsSrc.includes('ctx-emf-launcher'));
+  contextCardsSrc.includes('renderEMFAssessmentLauncher({ inModal: true') && contextCardSummariesSrc.includes('ctx-emf-launcher'));
 assert('45j. EMF launcher has dedicated context CSS',
   contextProfileCss.includes('.ctx-emf-launcher') &&
   contextProfileCss.includes('.ctx-emf-launcher-action') &&
